commit n°7

+++ Assets/Scripts/SpawnEnemy.cs
@@ -4,10 +4,10 @@ public class SpawnEnemy : MonoBehaviour
 {
     public GameObject[] enemies;
 
-    public float minPosX;
-    public float maxPosX;
-    public float minPosY;
-    public float maxPosY;
+    public int minPosX;
+    public int maxPosX;
+    public int minPosY;
+    public int maxPosY;
 
     public float delay;
     public float minDelay;
@@ -31,8 +31,8 @@ public class SpawnEnemy : MonoBehaviour
 
     public void SpawnEnemies()
     {
-        float randomPosX = Random.Range(minPosX, maxPosX);
-        float randomPosY = Random.Range(minPosY, maxPosY);
+        int randomPosX = Random.Range(minPosX, maxPosX);
+        int randomPosY = Random.Range(minPosY, maxPosY);
 
         delay = 0;
         Instantiate(enemies[Random.Range(0, enemies.Length)], new Vector3(randomPosX, randomPosY, 0), Quaternion.identity);
